Pull-up `saveAndLoad` test method
So that it can be reused by other NN tests.
