+++ src/test/java/at/medunigraz/imi/bst/n2c2/nn/LSTMEmbeddingsClassifierTest.java
@@ -3,6 +3,7 @@ package at.medunigraz.imi.bst.n2c2.nn;
 public class LSTMEmbeddingsClassifierTest extends BaseNNClassifierTest {
 
     public LSTMEmbeddingsClassifierTest() {
-        this.classifier = new LSTMEmbeddingsClassifier();
+        this.trainClassifier = new LSTMEmbeddingsClassifier();
+        this.testClassifier = new LSTMEmbeddingsClassifier();
     }
 }
